added openPgVectorStore besides createPgVectorStore
open will just open one and create the table if needed
create will really recreate the table (so drop it) before then opening
it again.

added embed(object, maxSegmentSizeInChars, maxOverlapSizeInChars) so
that the document is split first before embedding it

Search will now return multiply results for the same file. maybe we
should merge that and only take the most likely one (because getting the
text fragment back of a paragraph somewhere in the pdf has not much
meaning)